+++ Ladybug/Model/RadarMetadata.swift
@@ -17,10 +17,25 @@ struct RadarMetadata: Codable {
     let description: String
     let classification: String
     let originated: String
-    let product_version: String
+    let productVersion: String
     let number: String
     let user: String
     let id: Int
     let title: String
     let reproducible: String
+
+    enum CodingKeys: String, CodingKey {
+        case status
+        case resolved
+        case product
+        case description
+        case classification
+        case originated
+        case productVersion = "product_version"
+        case number
+        case user
+        case id
+        case title
+        case reproducible
+    }
 }