+++ LadybugTests/UserDefaultsExtensionTests.swift
@@ -18,6 +18,9 @@ class UserDefaultsExtensionTests: XCTestCase {
     
     override func tearDown() {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
+
+        UserDefaults.standard.radarOption = .openRadar
+        UserDefaults.standard.browserOption = .sfvcReader
         super.tearDown()
     }