+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Rewrite how to save Radars. Move to App Groups. Combine Radar, Radar Number and Radar Metadata
 - Refactor RadarNumber, URLParsers and Builders
 - Add tests
+- Backup & Restore as JSON
 
 ## [0.4(4)] - 2018-06-30
 ### Added